fix the buffer overflow in dynamic_bloom_test
Summary: int -> uint64_t

Test Plan:
it think it is pretty obvious
will run asan_check before committing

Reviewers: igor, haobo

Reviewed By: igor

CC: leveldb

Differential Revision: https://reviews.facebook.net/D17241

+++ util/dynamic_bloom_test.cc
@@ -69,7 +69,7 @@ static uint32_t NextNum(uint32_t num) {
 }
 
 TEST(DynamicBloomTest, VaryingLengths) {
-  char buffer[sizeof(int)];
+  char buffer[sizeof(uint64_t)];
 
   // Count number of filters that significantly exceed the false positive rate
   int mediocre_filters = 0;
